requirements.txt
# PyTorch (install with CUDA support: pip install torch torchvision torchaudio --index-url https://download.pytorch.org/whl/cu121)
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0

# Core dependencies
numpy>=1.20.0
scikit-learn>=1.0.0
pillow>=9.0.0
pandas>=1.3.0
scipy>=1.8.0

# Visualization and analysis
matplotlib>=3.5.0
seaborn>=0.12.0
tensorboard>=2.10.0

# Progress and utilities
tqdm>=4.60.0
warmup-scheduler>=0.3.0

# Jupyter environment (optional but recommended)
jupyterlab>=3.0.0
notebook>=6.4.0
ipykernel>=6.0.0

# Optional: For HuggingFace dataset access (if using prepare_wakevision.py)
# datasets>=2.0.0

# Optional: For ONNX export
# onnx>=1.12.0
# onnxruntime>=1.12.0 